Move resolver task implementations to base package
